Fix filename issue
If a url has a . after the last / and the text after the . is
not a valid extension, the file is saved with the incorrect
file extension. This fix use the content type of the response
to also save the file as json if the content type it is.

For example:

- http://url.json?limit=1 is now saved as .json?limit=1 , with the fix is
saved as .json

Signed-off-by: Yohanna Lisnichuk <yohanitalisnichuk@gmail.com>