Write out more info to disk - this is so when we try and restore files later, we have the info
For each file saved to disk, have an accompying info file

The important bit of info here to save is the data_type, to get this info we change the function call and so every scraper needs changes.

However, in doing so we reduce boiler plate which is good!

Also, pass encoding to process side properly.